rename: 패키지명 수정 (#63)
아니다 model이 낫겠다 죄송
