+++ backend/src/main/java/com/example/backend/domain/information/model/car/entity/ExteriorColor.java
@@ -1,14 +1,16 @@
-package com.example.backend.domain.information.entity.car.entity;
+package com.example.backend.domain.information.model.car.entity;
 
 import com.example.backend.domain.global.model.BaseInfo;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.relational.core.mapping.Embedded;
 import org.springframework.data.relational.core.mapping.Table;
 
-@Table("TRIM")
-public class Trim {
+@Table("EXTERIOR_COLOR")
+public class ExteriorColor {
     @Id
     private Long id;
+    private String colorCode;
     @Embedded(onEmpty = Embedded.OnEmpty.USE_EMPTY)
     private BaseInfo baseInfo;
+    private String comment;
 }
