+++ backend/src/main/java/com/example/backend/domain/information/model/car/entity/InteriorColor.java
@@ -1,4 +1,4 @@
-package com.example.backend.domain.information.entity.option.entity;
+package com.example.backend.domain.information.model.car.entity;
 
 import com.example.backend.domain.global.model.BaseInfo;
 import org.springframework.data.annotation.Id;
@@ -6,13 +6,13 @@ import org.springframework.data.jdbc.core.mapping.AggregateReference;
 import org.springframework.data.relational.core.mapping.Embedded;
 import org.springframework.data.relational.core.mapping.Table;
 
-@Table("ADDITIONAL_OPTION")
-public class AdditionalOption {
+@Table("INTERIOR_COLOR")
+public class InteriorColor {
     @Id
     private Long id;
-    private AggregateReference<AdditionalOption, Long> topOptionId;
-    private AggregateReference<Parts, Long> partsId;
-    @Embedded(onEmpty = Embedded.OnEmpty.USE_EMPTY)
+    private AggregateReference<ExteriorColor, Long> exteriorColorId;
+    private String iconSrc;
+    @Embedded(onEmpty = Embedded.OnEmpty.USE_NULL)
     private BaseInfo baseInfo;
-    private String detail;
+    private String comment;
 }
