+++ backend/src/main/java/com/example/backend/domain/information/model/car/entity/Trim.java
@@ -1,16 +1,14 @@
-package com.example.backend.domain.information.entity.option.entity;
+package com.example.backend.domain.information.model.car.entity;
 
 import com.example.backend.domain.global.model.BaseInfo;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.relational.core.mapping.Embedded;
 import org.springframework.data.relational.core.mapping.Table;
 
-@Table("BODYTYPE")
-public class Bodytype {
+@Table("TRIM")
+public class Trim {
     @Id
     private Long id;
     @Embedded(onEmpty = Embedded.OnEmpty.USE_EMPTY)
     private BaseInfo baseInfo;
-    private String detail;
-    private String comment;
 }
