+++ backend/src/main/java/com/example/backend/domain/information/model/option/entity/Bodytype.java
@@ -1,16 +1,16 @@
-package com.example.backend.domain.information.entity.car.entity;
+package com.example.backend.domain.information.model.option.entity;
 
 import com.example.backend.domain.global.model.BaseInfo;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.relational.core.mapping.Embedded;
 import org.springframework.data.relational.core.mapping.Table;
 
-@Table("EXTERIOR_COLOR")
-public class ExteriorColor {
+@Table("BODYTYPE")
+public class Bodytype {
     @Id
     private Long id;
-    private String colorCode;
     @Embedded(onEmpty = Embedded.OnEmpty.USE_EMPTY)
     private BaseInfo baseInfo;
+    private String detail;
     private String comment;
 }
