+++ backend/src/main/java/com/example/backend/domain/information/model/option/entity/DrivingSystem.java
@@ -1,15 +1,13 @@
-package com.example.backend.domain.information.entity.option.entity;
+package com.example.backend.domain.information.model.option.entity;
 
 import com.example.backend.domain.global.model.BaseInfo;
-import org.springframework.data.annotation.Id;
 import org.springframework.data.relational.core.mapping.Embedded;
 import org.springframework.data.relational.core.mapping.Table;
 
-@Table("POWERTRAIN")
-public class Powertrain {
-    @Id
+@Table("DRIVING_SYSTEM")
+public class DrivingSystem {
     private Long id;
-    @Embedded(onEmpty = Embedded.OnEmpty.USE_NULL)
+    @Embedded(onEmpty = Embedded.OnEmpty.USE_EMPTY)
     private BaseInfo baseInfo;
     private String detail;
     private String comment;
