+++ backend/src/main/java/com/example/backend/domain/sale/entity/SalesOption.java
@@ -1,6 +1,6 @@
 package com.example.backend.domain.sale.entity;
 
-import com.example.backend.domain.information.entity.option.entity.AdditionalOption;
+import com.example.backend.domain.information.model.option.entity.AdditionalOption;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.jdbc.core.mapping.AggregateReference;
 import org.springframework.data.relational.core.mapping.Table;
